Split test and add some for <img> with no src
